+++ Game/src/sys/CameraSystem.h
@@ -3,10 +3,10 @@
 class CameraSystem :
     public System
 {
-    int mBaseXOffset = 0;
-    int mBaseYOffset = 0;
+    float mBaseXOffset = 0;
+    float mBaseYOffset = 0;
 public:
-    CameraSystem(int x, int y, entt::registry& registry);
+    CameraSystem(float x, float y, entt::registry& registry);
     void update(double dt);
 };
